% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGenotypesHaploids.R
\name{sampleHaploids}
\alias{sampleHaploids}
\title{Simulate recombination events in R - output haplotypes}
\usage{
sampleHaploids(haploid.ids, haploid.parent, PLINK.prefix)
}
\arguments{
\item{haploid.ids}{vector of IDs to sample haploid genome from}

\item{haploid.parent}{parent who produced the haploid offspring, "MOTHER" or "FATHER"}

\item{PLINK.prefix}{default is a randomly generated name.}
}
\description{
This function simulates genotypes within a pedigree, based on centiMorgan map
distances. Outputs genotypes as a GenABEL gwaa object.
}
